requirements.txt
charmlibs-apt~=1.0
charmlibs-snap~=1.0
charmlibs-systemd~=1.0
ops~=3.2